etl-poc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.1
pydantic==2.8.2
pandas==2.2.2
pyarrow==16.1.0
fastavro==1.9.5
python-dotenv==1.0.1
